[REFACTOR] redis package 맞추기
